+++ app/ids-splitter/index.js
@@ -1,0 +1,1 @@
+import IdsSplitter from '../../src/ids-splitter';