+++ src/ids-splitter/ids-splitter.js
@@ -1,0 +1,66 @@
+import {
+  IdsElement,
+  customElement,
+  // attributes,
+  scss,
+  mix
+} from '../ids-base/ids-element';
+
+import {
+  IdsKeyboardMixin,
+  IdsEventsMixin,
+  IdsThemeMixin
+} from '../ids-mixins';
+
+import styles from './ids-splitter.scss';
+
+/**
+ * IDS Splitter Component
+ * @type {IdsSplitter}
+ * @inherits IdsElement
+ * @mixes IdsEventsMixin
+ * @mixes IdsThemeMixin
+ * @mixes IdsKeyboardMixin
+ * @part container - the container of all tabs
+ */
+@customElement('ids-splitter')
+@scss(styles)
+export default class IdsSplitter extends mix(IdsElement).with(
+    IdsEventsMixin,
+    IdsKeyboardMixin,
+    IdsThemeMixin
+  ) {
+  constructor() {
+    super();
+  }
+
+  /**
+   * Return the properties we handle as getters/setters
+   * @returns {Array} The properties in an array
+   */
+  static get attributes() {
+    return [];
+  }
+
+  /**
+   * Create the Template to render
+   *
+   * @returns {string} the template to render
+   */
+  template() {
+    return (
+      `<div
+        class="ids-splitter"
+        part="splitter"
+      ></div>`
+    );
+  }
+
+  connectedCallback() {
+    super.connectedCallback?.();
+  }
+
+  disconnectedCallback() {
+    super.disconnectedCallback?.();
+  }
+}