+++ doc/CHANGELOG.md
@@ -114,6 +114,8 @@
   - Markup has changed to a custom element `<ids-spinbox></ids-spinbox>`
   - If using events, events are now plain JS events.
   - Can now be imported as a single JS file and used with encapsulated styles
+- `[Splitter]` The splitter component has been changed to a web component and renamed to ids-splitter.
+  - TODO: DETAILS
 - `[Switch]` The Switch component has been changed to a web component and renamed to ids-switch.
   - Markup has changed to a custom element `<ids-switch></ids-switch>`
   - If using events, events are now plain JS events.